Code refactoring: add SkoutAppletPanelExtension, make SkoutPanel a singleton
SkoutPanel is now a singleton because I don't plan to add multipanel support
(yet?). Kicker does that better anyway.

SkoutAppletPanelExtension is a class which exposes some methods of the panel
to applets without having to include the whole panel sources in libskout.
It is a singleton accessible from all applets via the panelExt static variable.
Its methods emit signals which the panel is expected to connect to.

As a result, we don't pollute SkoutApplet with methods and they are globally
accessible (previously you had to go up the parent hierarchy until the applet
itself in order to invoke a method).

Signed-off-by: Mavridis Philippe <mavridisf@gmail.com>