Better autostart implementation
This is a more elegant solution based on desktop files installed
in special autostart directories.

If Skout is to run alongside Kicker, its own desktop file is just
installed into user's $TDEHOME/Autostart.

If Skout is to replace Kicker, it writes a special desktop file
to $TDEHOME/share/autostart/panel.desktop. This file is read by
tdeinit and executed on session start, and we can obviously
override it on a per-user basis.

For the second scenario we no longer need to hook into the
system configuration loading process and manually kill Kicker
after it has already started. Instead, Skout is launched
immediately, resulting in a more pleasant visual experience.

Signed-off-by: Philippe Mavridis <philippe.mavridis@yandex.com>